Store and use userID after successful login/registration
The userID is now stored globally after a user logs in or registers.
This userID is then used when fetching data for the main table.

After successful registration, the user is now navigated to the main table instead of the login page.
A snackbar with a success message from the API is also displayed.